pyproject.toml
[project]
name = "trace-algonauts"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "datalad>=1.3.2",
    "h5py>=3.16.0",
    "ipykernel>=7.2.0",
    "ipywidgets>=8.1.8",
    "matplotlib>=3.10.8",
    "nibabel>=5.4.2",
    "nilearn>=0.13.1",
    "numpy>=2.4.3",
    "pandas>=3.0.1",
    "pyyaml>=6.0.3",
    "seaborn>=0.13.2",
    "umap-learn>=0.5.12",
]
